fix(coap): Update to libcoap v4.3.5a
Fixes some critical issues.
