+++ coap/examples/coap_client/main/coap_client_example_main.c
@@ -31,6 +31,7 @@
 
 #include "protocol_examples_common.h"
 
+#include "coap_config.h"
 #include "coap3/coap.h"
