+++ coap/examples/coap_server/main/coap_server_example_main.c
@@ -29,6 +29,7 @@
 
 #include "protocol_examples_common.h"
 
+#include "coap_config.h"
 #include "coap3/coap.h"
 
 #ifndef CONFIG_COAP_SERVER_SUPPORT
