+++ coap/examples/coap_server/main/idf_component.yml
@@ -1,8 +1,8 @@
-version: "1.0.0"
+version: 1.0.0
 description: CoAP Server Example
 dependencies:
   espressif/coap:
-    version: "^4.3.0"
-    override_path: '../../../'
+    version: ^4.3.0
+    override_path: ../../../
   protocol_examples_common:
     path: ${IDF_PATH}/examples/common_components/protocol_examples_common
