+++ coap/idf_component.yml
@@ -1,4 +1,4 @@
-version: "4.3.5~3"
+version: "4.3.5~4"
 description: Constrained Application Protocol (CoAP) C Library
 url: https://github.com/espressif/idf-extra-components/tree/master/coap
 dependencies:
