+++ coap/sbom_libcoap.yml
@@ -4,7 +4,7 @@ cpe: cpe:2.3:a:libcoap:libcoap:{}:*:*:*:*:*:*:*
 supplier: 'Organization: libcoap <https://libcoap.net/>'
 description: A CoAP (RFC 7252) implementation in C
 url: https://github.com/obgm/libcoap
-hash: 4d84859620540c8baf06fa094abb9e0ae19da0e7
+hash: e3fdcdcfbd1588754fe9dd4b754ac9397260f0f9
 cve-exclude-list:
   - cve: CVE-2024-31031
     reason: Resolved in version 4.3.5-rc1
@@ -14,3 +14,23 @@ cve-exclude-list:
     reason: Resolved in version 4.3.5-rc3
   - cve: CVE-2025-50518
     reason: Not applicable as per comment https://github.com/obgm/libcoap/issues/1724#issuecomment-3296780541
+  - cve: CVE-2025-59391
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69493
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69494
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69495
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69496
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69497
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69498
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69499
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69500
+    reason: Resolved in version 4.3.5a
+  - cve: CVE-2025-69501
+    reason: Resolved in version 4.3.5a
